+++ tests/integration/test_charm.py
@@ -4,18 +4,24 @@
 
 
 import logging
-from typing import List
 
 import psycopg2
 import pytest
 import requests
 from pytest_operator.plugin import OpsTest
 
-from tests.helpers import METADATA, STORAGE_PATH
+from tests.helpers import STORAGE_PATH
+from tests.integration.helpers import (
+    APP_NAME,
+    build_application_name,
+    convert_records_to_dict,
+    db_connect,
+    get_postgres_password,
+    get_primary,
+)
 
 logger = logging.getLogger(__name__)
 
-APP_NAME = METADATA["name"]
 SERIES = ["focal"]
 
 
@@ -125,23 +131,19 @@ async def test_persist_data_through_graceful_restart(ops_test: OpsTest, series: 
     # Set a composite application name in order to test in more than one series at the same time.
     application_name = build_application_name(series)
 
-    primary = await get_primary(ops_test)
-    password = await get_postgres_password(ops_test)
-    for unit in ops_test.model.applications[APP_NAME].units:
-        if unit.name == primary:
-            address = unit.public_address
-        else:
-            replica = unit.name
+    primary = await get_primary(ops_test, f"{application_name}/0")
+    password = await get_postgres_password(ops_test, primary)
 
     # Write data to primary IP.
-    logger.info(f"connecting to primary {primary} on {address}")
-    with db_connect(host=address, password=password) as connection:
+    host = ops_test.model.units.get(primary).public_address
+    logger.info(f"connecting to primary {primary} on {host}")
+    with db_connect(host=host, password=password) as connection:
         connection.autocommit = True
         connection.cursor().execute("CREATE TABLE gracetest (testcol INT );")
 
     # Remove one unit.
     await ops_test.model.destroy_units(
-        replica,
+        primary,
     )
     await ops_test.model.wait_for_idle(apps=[application_name], status="active", timeout=1000)
 
@@ -150,61 +152,9 @@ async def test_persist_data_through_graceful_restart(ops_test: OpsTest, series: 
     await ops_test.model.wait_for_idle(apps=[application_name], status="active", timeout=1000)
 
     # Testing write occurred to every postgres instance by reading from them
-    for unit in ops_test.model.applications[APP_NAME].units:
+    for unit in ops_test.model.applications[application_name].units:
         host = unit.public_address
         logger.info("connecting to the database host: %s", host)
         with db_connect(host=host, password=password) as connection:
             # Ensure we can read from "gracetest" table
             connection.cursor().execute("SELECT * FROM gracetest;")
-
-
-def build_application_name(series: str) -> str:
-    """Return a composite application name combining application name and series."""
-    return f"{APP_NAME}-{series}"
-
-
-def convert_records_to_dict(records: List[tuple]) -> dict:
-    """Converts pyscopg2 records list to a dict."""
-    dict = {}
-    for record in records:
-        # Add record tuple data to dict.
-        dict[record[0]] = record[1]
-    return dict
-
-
-async def get_primary(ops_test: OpsTest, unit_id=0) -> str:
-    """Get the primary unit.
-
-    Args:
-        ops_test: ops_test instance.
-        unit_id: the number of the unit.
-
-    Returns:
-        the current primary unit.
-    """
-    action = await ops_test.model.units.get(f"{APP_NAME}/{unit_id}").run_action("get-primary")
-    action = await action.wait()
-    return action.results["primary"]
-
-
-async def get_postgres_password(ops_test: OpsTest):
-    """Retrieve the postgres user password using the action."""
-    unit = ops_test.model.units.get(f"{APP_NAME}/0")
-    action = await unit.run_action("get-initial-password")
-    result = await action.wait()
-    return result.results["postgres-password"]
-
-
-def db_connect(host: str, password: str):
-    """Returns psycopg2 connection object linked to postgres db in the given host.
-
-    Args:
-        host: the IP of the postgres host
-        password: postgres password
-
-    Returns:
-        psycopg2 connection object linked to postgres db, under "postgres" user.
-    """
-    return psycopg2.connect(
-        f"dbname='postgres' user='postgres' host='{host}' password='{password}' connect_timeout=10"
-    )
